backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
langchain==0.3.9
langgraph==0.2.53
langchain-openai==0.2.9
langchain-core==0.3.21
python-dotenv==1.0.1
pydantic==2.10.3
requests==2.32.3
gspread==6.1.2
oauth2client==4.1.3